move to RetirementHandler and C3-Offset contract
